+++ src/Organisms/_organisms.story.js
@@ -1,3 +1,3 @@
 import './Form/_form.story';
-import './Form.next/_form.story';
+import './DynamicForm/_form.story';
 import './Table/_table.story';
